fix: keep home as format chooser and stop /login redirect loop
- Home (/) is the tournament-format landing page again (create-with-name
  flow), instead of redirecting to /tournaments.
- Middleware no longer redirects authenticated users away from /login|/signup;
  a stale/invalid cookie (localhost cookies are shared across ports) was
  bouncing /login -> home while the server bounced back -> ERR_TOO_MANY_REDIRECTS.
  Genuinely signed-in users are now redirected off the auth pages client-side
  via a validated session.
- Remove /tournaments/new (its content is the home page now); point "New
  Tournament" links and the header logo at /.
